fix(execution): sign-extend the 16-bit register bit index
BitExecutor::index sign-extends the bit index for the 32- and 64-bit forms but
zero-extended the 16-bit one. Against a memory bit base the index is not masked at
all, it is shifted right by three to pick the containing byte, so a negative 16-bit
index addressed far above the operand instead of below it: `bt %ax,(%rsi)` with ax
= -8 read at rsi + 0x1fff rather than rsi - 1, which faulted here rather than
returning a wrong bit only because the operand happened to be near the end of its
mapping.

The lowering was already correct: hl_x86_emit_address sign-extends with sxth for a
16-bit index and sxtw for a 32-bit one before the arithmetic shift, so this was the
native and interpreted paths disagreeing with the interpreter wrong. qemu-x86_64
confirms the signed reading, placing bit -8 in the byte below the base.

Found by sweeping the register-index memory form of bt/bts/btr/btc over indices
either side of zero and either side of the byte and word boundaries, which is the
non-immediate half of the form whose immediate half carried the original mask bug.

+++ src/runtime/hl-execution/src/x86/bit_operation.rs
@@ -162,7 +162,7 @@ impl BitExecutor {
     fn index(cpu: &CpuState, operand: ScalarOperand, width: ScalarWidth) -> i64 {
         match operand {
             ScalarOperand::Register(register) => match width {
-                ScalarWidth::Word => i64::from(cpu.read_register(register, width) as u16),
+                ScalarWidth::Word => i64::from(cpu.read_register(register, width) as u16 as i16),
                 ScalarWidth::Dword => cpu.read_register(register, width) as i32 as i64,
                 ScalarWidth::Qword => cpu.read_register(register, width) as i64,
                 ScalarWidth::Byte => unreachable!(),
